Add Calculator demo
This builds a demo application that builds libjson-c from its git
repository, a libcalc.so from some c files and links everything together
into an executable.

This tests in addition to compile and linking with gcc library lookup
via -L and LIBRARY_PATH environment variable.

It also test that multiple build jobs and repeated builds of the same
files are handled properly by the database. Therefore the "bug" in the
demo's Makefile that causes the four sourcefiles to be build 4 times is
done on purpose and should not be "fixed".

To run the build script a QMSTR_HOME environment variable must be set to
a directory that contains a bin directory containing at least a gcc
symlinked to qmstr-wrapper.