Create demo cases in container (#12)
* Create container for our demo cases

We have the calc demo. Reporting is not yet included.

* Move demos to their own demos folder

* Set demos as Volume

Set demos as Volume so qmstr.yaml configuration file will be passed to master container

* Call reporting's setup.sh

* Call the correct entrypoint

* Fix bug

* add qmstr to demo container for reporting

* Change path for qmstr-master, call report and analyse using ENVVAR

* Add curl demo case

* Remove clean up method

We have already implemented the removal of the master container in build.inc

* Move curl to base directory

Curl is downloaded in the directory of build.inc

* Rebase to create demo docker commit

* Create container for our demo cases

* Set demos as Volume

Set demos as Volume so qmstr.yaml configuration file will be passed to master container

* Call reporting's setup.sh

* Call the correct entrypoint

* add qmstr to demo container for reporting

* Change path for qmstr-master, call report and analyse using ENVVAR

* Add curl demo case

* Remove clean up method

We have already implemented the removal of the master container in build.inc

* Move curl to base directory

Curl is downloaded in the directory of build.inc

* fix # at bash call

* trigger master exit even if script fails + call reporting in curl demo

* Copy qmstr dir in demo container

We need the path for the reporter

* Remove duplicated copies of binaries

We already have these binaries from runtime.